CMPE-220: HW4: Enlarge the memory size and stack pointer
Since we are going to run a larger program that needs more memory space
for both instruction memory and data memory. This patch just enlarge them.

The only drawback of this operation and the reason why I didn't make them
large enough at the first place is that the output of the highest debug
level will be too long to trace useful information.

Signed-off-by: Nicolin Chen <nicoleotsuka@gmail.com>